% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_partitioned_prevalence.R
\name{age_partitioned_prevalence}
\alias{age_partitioned_prevalence}
\title{age_partitioned_prvalence}
\usage{
age_partitioned_prevalence(overall_prevalence1, overall_prevalence2,
  sigma_prevalence, cluster_number, cluster_size, ages)
}
\arguments{
\item{overall_prevalence1}{is a vector of age specific prevalence  at time t1}

\item{overall_prevalence2}{is a vector of age specific prevalence  at time t2}

\item{sigma_prevalence}{is a vector with the  standard errors associated with age specific prevalence  at time t1}

\item{cluster_number}{unique id number of the cluster}

\item{cluster_size}{sizes of the clusters}

\item{ages}{the ages being sampled}
}
\value{
returns an array of dimensions time t , age a and time since infection - tau
}
\description{
a functions that partitions realised prevalence per age into the regions of a stipulated.
}
