% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_time_structured_population.R
\name{age_time_structured_population}
\alias{age_time_structured_population}
\title{age_time_structured_population}
\usage{
age_time_structured_population(time_slice, max_birth_date, min_birth_date,
  time_step, max_age, type, birth_rate, probability_of_recent_infection,
  base_mortality_function, pmtct_birth_rate, incidence_function,
  excess_mortality_function)
}
\arguments{
\item{time_slice}{dates the surveys are to be conducted}

\item{max_birth_date}{minimum date of birth of the cohorts}

\item{min_birth_date}{minimum date of birth of the cohorts}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts.}

\item{max_age}{the maximum age of each cohort}

\item{type}{the type of function to be used in estimating the probability of testing recently infected}

\item{birth_rate}{bith rate at the the give date}

\item{probability_of_recent_infection}{function in use in calculating  probability of being recently infected}

\item{base_mortality_function}{base mortality rate ata a given age and time}

\item{pmtct_birth_rate}{rate of babies/proportion of the population born HIV positive}

\item{incidence_function}{Incidence rate at a given time}

\item{excess_mortality_function}{excess  mortality rate ata a given age and time}
}
\description{
age_time_structured_population
}
