% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregated_survey.R
\name{aggregated_survey}
\alias{aggregated_survey}
\title{aggregated_survey}
\usage{
aggregated_survey(survey_specs, prevalencesdata, time_step,
  reporting_bin = 1)
}
\arguments{
\item{survey_specs}{demographic specifications of the survey.}

\item{prevalencesdata}{output from the simulation with ages, HIV prevalence, and prevalence of recency.}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts}

\item{reporting_bin}{reporting age bin in a time slice}
}
\value{
individual level data with age, HIV status, and HIV recency status.
}
\description{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
