% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mortality_matrix.R
\name{base_mortality_vector}
\alias{base_mortality_vector}
\title{base_mortality_vector}
\usage{
base_mortality_vector(max_age, list_of_times, base_mortality, time_step)
}
\arguments{
\item{max_age}{The maximum age attained by each birth cohort (aging steps).}

\item{list_of_times}{a numeric vectors indicating the dates in simulation. Note that date format is not used.}

\item{base_mortality}{a function which takes as arguments age and time and returns a numberic rate of mortality for each age and time included in the simulation.
This function can be defined by user or can be selected from several default options included in the package.
The user-defined or package default function should be called by name when included as an argument in the base_mortality_matrix function}

\item{time_step}{the time step between consecurtive birth dates in the vector list_of_birth_times}
}
\value{
a matrix of column length max_age and row length list_of_birth_times,
Values stored in the matrix are numeric double, from 0-1, which represent the natural mortality rate at given age and time
}
\description{
a function that returns a matrix of probabilities of mortality for each ages and time step of the simulation
}
