% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_counts.R
\name{birth_counts}
\alias{birth_counts}
\title{birth_counts_fun}
\usage{
birth_counts(dates_needing_birth_counts, birth_rate, time_step)
}
\arguments{
\item{dates_needing_birth_counts}{a vector of births dates to be assigned the birth count}

\item{birth_rate}{a function  for the distribution of births in the interval}

\item{time_step}{the time step between consecurtive birth dates in the vector list_of_birth_times}
}
\value{
Returns a vector of bith counts
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
