% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_counts_fun.R
\name{birth_counts_fun}
\alias{birth_counts_fun}
\title{birth_counts_fun}
\usage{
birth_counts_fun(t_1, t_2, birthrate, delta)
}
\arguments{
\item{t_1}{start time of the simulation}

\item{t_2}{end time of the simulation}

\item{birthrate}{a function  for the distribution of births in interval}

\item{delta}{the time step between consecurtive birth_dates}
}
\value{
Returns a vector of bith counts  from the inputs total_births, delta and birth_dates of length min-birth_dates:max-birth_dates
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
