% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_rate_fun.R
\name{birth_rate_fun}
\alias{birth_rate_fun}
\title{birth_rate_fun}
\usage{
birth_rate_fun(birth_dates, delta)
}
\arguments{
\item{birth_dates}{the calender dates of births.}

\item{delta}{the time step between consecurtive birth_dates}
}
\value{
Returns a vector of bith counts  from the inputs total_births, delta and birth_dates of length min-birth_dates:max-birth_dates
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
\examples{
generate_birth_counts(1000, 1984:1990, 1)
generate_birth_counts(10000, 2005 : 2018, 1)

birth_count <- birth_counts (total_births = 1000,
 birth_dates = 1945:1950,
 delta = 1)

}
