% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_of_recently_infected.R
\name{calculate_MDRI}
\alias{calculate_MDRI}
\title{calculate_MDRI}
\usage{
calculate_MDRI(function_prt = probability_of_recently_infected,
  big_Tyears)
}
\arguments{
\item{function_prt}{dates the surveys are to be conducted}

\item{big_Tyears}{the maximum age of each cohort}
}
\description{
calculate_MDRI
}
