% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_prevalence.R
\name{calculate_prevalence}
\alias{calculate_prevalence}
\title{calculate_prevalence}
\usage{
calculate_prevalence(population_at_date, infection_time = NULL)
}
\arguments{
\item{population_at_date}{a matrix of age and times since infection at a specified calender date.}

\item{infection_time}{specifies the output of interest wether or not its the total prevalence per age group or
prevalence per specified age and time since infection.
total prevalence for all ranges at the current date.}
}
\value{
prevalences for ages at given current date.
}
\description{
A function that extracts prevalence by age and time since infection
}
