% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortcounts.R
\name{cohortcounts}
\alias{cohortcounts}
\title{cohortcounts}
\usage{
cohortcounts(cohort, date_of_birth, time_slice)
}
\arguments{
\item{cohort}{cohort history at survey timeslices.}

\item{date_of_birth}{date of birth of the cohort}

\item{time_slice}{times the cross sectional surveys are conducted}
}
\value{
a long data frame with ages and population counts at a specified time.
}
\description{
calculates the probability of being recently infected given that you already HIV positive.
}
\details{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
