% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_populationarray.R
\name{collapse_populationarray}
\alias{collapse_populationarray}
\title{collapse_populationarray}
\usage{
collapse_populationarray(data_set, list_of_birth_times, max_age, time_step,
  counts = FALSE)
}
\arguments{
\item{data_set}{the minimum date of birth for the birth cohorts}

\item{list_of_birth_times}{the maximum date of birth for the birth cohorts}

\item{max_age}{maximum age attained by each birth cohort}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts}

\item{counts}{takes in either true or false if true the function returns population counts, otherwise prevalence.}
}
\description{
A wrapper function that returns a list of the the susceptible and infected population
}
