% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compact_population.R
\name{compact_population}
\alias{compact_population}
\title{compact_population}
\usage{
compact_population(susceptible, infected)
}
\arguments{
\item{susceptible}{proportion of the population that is susceptible.}

\item{infected}{proportion of the population that is infected.}
}
\value{
a array

Values stored in the matrix are numeric double, from 0-1, which represent the natural mortality rate at given age and time
}
\description{
a function that returns an array
}
