% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_age_time_matrix.R
\name{compress_age_time_matrix}
\alias{compress_age_time_matrix}
\title{compress_age_time_matrix}
\usage{
compress_age_time_matrix(age_time_matrix)
}
\arguments{
\item{age_time_matrix}{the matrix to be transformed}
}
\value{
returns a matrix with a rectangular format
}
\description{
a function that takes in a matrix with a pararellogram format and transforms it into a rectangular format and assigns column and row names.
}
