% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_indept_age_tent_incidence.R
\name{constant_incidence}
\alias{constant_incidence}
\title{constant_incidence}
\usage{
constant_incidence(ages, times)
}
\arguments{
\item{ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{times}{numeric, indicates time or times at which the incidence rate is desired}
}
\description{
a function that takes as arguments age and time and returns a numeric vector of length 1
representing a rate of incidence at the indicated age and time
The generate_incidence function is required as an argument for the do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
