% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constant_pmtct_rate.R
\name{constant_pmtct_rate}
\alias{constant_pmtct_rate}
\title{constant_pmtct_rate}
\usage{
constant_pmtct_rate(dates_as_float, constant_value = 0)
}
\arguments{
\item{dates_as_float}{is a double/vector of times the birth rate is to be evaluated at.}

\item{constant_value}{the number of births at a given birth date, given that it is constant through out the specified birth dates}
}
\value{
Returns a vector of birth counts  of the length of the birth dates
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
