% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_probability_surviving_infected.R
\name{cumulative_probability_surviving_infected}
\alias{cumulative_probability_surviving_infected}
\title{cumulative_probability_surviving_infected}
\usage{

  cumulative_probability_surviving_infected(probability_surviving_infected_array)
}
\arguments{
\item{probability_surviving_infected_array}{3 dimensional probability matrix}
}
\value{
returns an array of dimensions time, age and time since infection - tau
}
\description{
a function that returns a matrix of probabilities of mortality for each age and time step of the simulation
}
