% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_one_simulation.R
\name{do_one_simulation}
\alias{do_one_simulation}
\title{do_one_simulation}
\usage{
do_one_simulation(first_birth_time, last_birth_time, time_step, max_age,
  birth_rate, base_mortality, incidence, excess_mortality)
}
\arguments{
\item{first_birth_time}{the minimum date of birth for the birth cohorts}

\item{last_birth_time}{the maximum date of birth for the birth cohorts}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts}

\item{max_age}{maximum age attained by each birth cohort}

\item{birth_rate}{the birth rate in the hypothetical population at the specified times}

\item{base_mortality}{a function that specifies  the rate of occurence of natural deaths with arguments age and time.}

\item{incidence}{a function that specifies  the rate of occurence of the infections with arguments age and time.}

\item{excess_mortality}{a function that specifies  the rate of occurence ofdisease induced deaths with arguments age and time.}
}
\value{
a matrix of column length max_age and row length list_of_birth_times,
Values stored in the matrix are numeric double, from 0-1, which represent the probability of becoming infected at age and time
}
\description{
A wrapper function that returns a list of the the susceptible and infected population
}
