% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excess_mortality.R
\name{excess_mortality}
\alias{excess_mortality}
\title{excess_mortality}
\usage{
excess_mortality(matrix_of_age, matrix_of_time, time_since_i,
  constant = 0, shape_parameter = 2, median_survival = 10.6)
}
\arguments{
\item{matrix_of_age}{numeric, indicates age}

\item{matrix_of_time}{numeric, indicates average time}

\item{time_since_i}{umeric, indicates time since infection among the infected poplation}

\item{constant}{numeric, indicates a constant rate of excess mortality}

\item{shape_parameter}{numeric, indicates weibull shape parameter}

\item{median_survival}{numeric, indicates median survival time of the infected individuals}
}
\value{
a numeric vector that represents the excess mortality rate at age, time since infection and time.
}
\description{
a function that takes as arguments age,time and tau - which indicates the average time since infection among the infected population - and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of excess mortality -i.e. among infected population relative to non-infected population - at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
