% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extactcohorttau_distribution.R
\name{extactcohorttau_distribution}
\alias{extactcohorttau_distribution}
\title{birth_cohort_simulation}
\usage{
extactcohorttau_distribution(birth_date, population)
}
\arguments{
\item{birth_date}{the date of birth of the cohort}

\item{population}{the minimum date of birth for the birth cohorts}
}
\value{
a matrix of column length max_age and row length list_of_birth_times,
Values stored in the matrix are numeric double, from 0-1, which represent the probability of becoming infected at age and time
}
\description{
A wrapper function that returns a list of the the susceptible and infected population
}
