% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extracting_population_status.R
\name{extract_population_status}
\alias{extract_population_status}
\title{extract_population_status}
\usage{
extract_population_status(survey_date, first_birth_date, time_step,
  population)
}
\arguments{
\item{survey_date}{date at which prevalences are required}

\item{first_birth_date}{the minimum date of birth for the birth cohorts}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts}

\item{population}{an array with the population history.}
}
\value{
prevalences for ages at given current date.
}
\description{
A function that extracts the population status at the specified survey date
}
