% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_base_mortality.R
\name{generate_base_mortality}
\alias{generate_base_mortality}
\title{generate_base_mortality}
\usage{
generate_base_mortality(t, constant = 0.01, age_min = 1,
  age_max = 50, exmin = 0, exfin = 0.01)
}
\arguments{
\item{t}{numeric, indicates time or times at which the mortality rate is desired}

\item{constant}{numeric, indicates a constant rate of mortality}

\item{age_min}{numeric, indicates minimum age to be included in the simulation}

\item{age_max}{numeric, indicates maximum age to be included in the simulation}

\item{exmin}{numeric, indicates minimum mortality, which is at age_min}

\item{exfin}{numeric, indicates maximum/final mortality at age_max, unless otherwise specified by user defined function}
}
\value{
a numeric vector that represents the mortality rate at t.
}
\description{
a function that takes as arguments age and time and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of mortality at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
