% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_birth_counts.R
\name{generate_birth_counts}
\alias{generate_birth_counts}
\title{generate_birth_counts}
\usage{
generate_birth_counts(total_births, birth_dates, delta)
}
\arguments{
\item{total_births}{the total number of births between maximum and minimun birth_dates given.}

\item{birth_dates}{the calender dates of births.}

\item{delta}{the time step between consecurtive birth_dates}
}
\value{
Returns a vector of bith counts  from the inputs total_births, delta and birth_dates of length min-birth_dates:max-birth_dates
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
\examples{
generate_birth_counts(1000, 1984:1990, 1)
generate_birth_counts(10000, 2005 : 2018, 1)
}
