% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_excess_mortality_tau.R
\name{generate_excess_mortality_tau}
\alias{generate_excess_mortality_tau}
\title{generate_excess_mortality_tau}
\usage{
generate_excess_mortality_tau(t, ta, constant = 0.05, age_min = 0,
  age_max = 50, exmin = 0.01, exfin = 0.05)
}
\arguments{
\item{t}{numeric, indicates time or times at which the excess mortality rate is desired}

\item{ta}{numeric, indicates average time since infection among the infected poplation}

\item{constant}{numeric, indicates a constant rate of excess mortality}

\item{age_min}{numeric, indicates minimum age to be included in the simulation}

\item{age_max}{numeric, indicates maximum age to be included in the simulation}

\item{exmin}{numeric, indicates minimum excess mortality, which is at age_min}

\item{exfin}{numeric, indicates maximum/final excess mortality at age_max, unless otherwise specified by user defined function}
}
\value{
a numeric vector that represents the excess mortality rate at age and time.
}
\description{
a function that takes as arguments age,time and tau - which indicates the average time since infection among the infected population - and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of excess mortality -i.e. among infected population relative to non-infected population - at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
