% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_incidence.R
\name{generate_incidence}
\alias{generate_incidence}
\title{generate_incidence}
\usage{
generate_incidence(t, constant = 0.05, age_min = 0, age_max = 50,
  age_peak = 25, Imin = 0.01, Ipeak = 0.05, Ifin = 0.02)
}
\arguments{
\item{t}{numeric, indicates time or times at which the incidence rate is desired}

\item{constant}{numeric, indicates a constant rate of incidence}

\item{age_min}{numeric, indicates minimum age to be included in the simulation}

\item{age_max}{numeric, indicates maximum age to be included in the simulation}

\item{age_peak}{numberic, indicates age at which indicence reaches maximum, which is at Ipeak}

\item{Imin}{numeric, indicates minimum incidence, which is at age_min}

\item{Ipeak}{numeric, indicates maximum incidence at Ipeak}

\item{Ifin}{numeric, indicates incidence at Ipeak}
}
\value{
a numeric vector that represents the incidence rate at age and time.
}
\description{
a function that takes as arguments age and time and returns a numeric vector of length 1
representing a rate of incidence at the indicated age and time
The generate_incidence function is required as an argument for the do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
