% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_infected_tau_0.R
\name{generate_infected_tau_0}
\alias{generate_infected_tau_0}
\title{generate_infected_tau}
\usage{
generate_infected_tau_0(incidence_matrix, susceptible_pop_counts)
}
\arguments{
\item{incidence_matrix}{a matrix of susceptible population counts, as defined by the generate_susceptibles function}

\item{susceptible_pop_counts}{a matrix of incidence, as defined by the generate_incidence_matrix function}
}
\value{
a matrix of row length age_steps and column length birth_dates. Numeric values in each cell of the matrix
represent the number of infected individuals in the population whose time since infection is 0
}
\description{
generate_infected_tau_0 returns a matrix of numeric values for age and time, representing the number of individuals within the population whose time since infection is 0
a matrix of susceptible population counts - as defined by the generate_susceptibles function - and...
a matrix of incidence - as defined by the generate_incidence_matrix function
}
\examples{
x <- generate_infected_tau_0 (incidence_matrix = matrix(rep(0.04, 4), 2, 2),
susceptible_pop_counts = matrix(rep(100, 4), 2, 2))
}
