% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_susceptible_cumulative_survival.R
\name{generate_susceptible_cumulative_survival_matrix}
\alias{generate_susceptible_cumulative_survival_matrix}
\title{generate_susceptible_cumulative_survival}
\usage{

  generate_susceptible_cumulative_survival_matrix(susceptible_survival_rate_matrix)
}
\arguments{
\item{susceptible_survival_rate_matrix}{a survival probability matrix, as defined by the generate_susceptible_surv_rate function}
}
\value{
a matrix of column length age_steps and row length birth_dates, calculated from the susceptible_survival_rate_matrix
Values stored in the matrix are numeric-double, from 0-1, which represent the cumulative probability of not dying and not aquiring infection among the susceptible population
}
\description{
a function that returns a matrix of cumulative probabilities of survival -i.e. not dying and aquiring infection - for each age and time step of the simulation
}
\examples{
x <- generate_susceptible_cumulative_survival_matrix
(matrix(seq(0.01, 0.16, 0.01), ncol = 2 , nrow = 8))
}
