% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_susceptible_survival_rate.R
\name{generate_susceptible_surv_rate}
\alias{generate_susceptible_surv_rate}
\title{generate_susceptible_survival_rate}
\usage{
generate_susceptible_surv_rate(incidence_matrix, base_mortality_matrix)
}
\arguments{
\item{incidence_matrix}{probability matrix of getting infected in the susceptibles}

\item{base_mortality_matrix}{probability matrix of getting infected in the sinfected}
}
\value{
returns a probability of not getting infected or dying within  susceptibles from the incidence_matrix and base_mortality_matrix
}
\description{
a function that calculates the survival probabilities in the susceptibles
}
\examples{
x <- generate_susceptible_surv_rate (matrix(seq(0.01, 0.16, 0.01),
ncol = 2 , nrow = 8),
matrix(seq(0.01, 0.16, 0.01),
ncol = 2 , nrow = 8))
}
