% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_susceptibles.R
\name{generate_susceptibles}
\alias{generate_susceptibles}
\title{generate_susceptibles}
\usage{
generate_susceptibles(survival_matrix, birth_counts)
}
\arguments{
\item{survival_matrix}{matrix of survival probabilities for each age and time, defined by package's generate_susceptible_cumulative_survival_matrix function.}

\item{birth_counts}{vector of numberic values representing the number of  birth_counts at age and time}
}
\value{
a matrix of row length age_steps and column length birth_dates. Numeric values in each cell of the matrix represent the number of individuals in the population who are alive and not infected
}
\description{
a function that returns a matrix of numeric values for age and time, representing the number of individuals within the population who are alive and not infected
}
