% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_indept_age_tent_incidence.R
\name{incidence_mahiane}
\alias{incidence_mahiane}
\title{incidence_mahiane}
\usage{
incidence_mahiane(matrix_of_ages, matrix_of_times, age_debut = 0,
  beta = 2.3, sigm2 = 0.5)
}
\arguments{
\item{matrix_of_ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{matrix_of_times}{numeric, indicates time or times at which the incidence rate is desired}

\item{age_debut}{age at which individuals get HIV infected}

\item{beta}{shape parameter for the log normal distribution}

\item{sigm2}{scale parameter for the log normal distribution}
}
\value{
a numeric vector that represents the incidence rate at age and time.
}
\description{
a function that takes as arguments age and time and returns a numeric vector of length 1
representing a rate of incidence at the indicated age and time
The generate_incidence function is required as an argument for the do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
