% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_exp.R
\name{incidence_matrix_exp}
\alias{incidence_matrix_exp}
\title{incidence_matrix_fun}
\usage{
incidence_matrix_exp(incidence_matrix, base_mortality_matrix, delta)
}
\arguments{
\item{incidence_matrix}{an incidence matrix}

\item{base_mortality_matrix}{a mortality matrix}

\item{delta}{the time step between consecurtive birth_dates}
}
\value{
a matrix of the probabiity of getting infected at a given age and time.
Values stored in the matrix are numeric double, from 0-1, which represent the probability of becoming infected at age and time
}
\description{
a function that returns a matrix of probabilities for each age and time step of the simulation
}
