% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incidence_matrix_fun.R
\name{incidence_matrix_fun}
\alias{incidence_matrix_fun}
\title{incidence_matrix_fun}
\usage{
incidence_matrix_fun(age_steps, birth_dates, generate_incidence_fun, delta)
}
\arguments{
\item{age_steps}{the number of steps to age the population}

\item{birth_dates}{a numeric vectors of length min:max; indicates the range of ages to be included in simulation. Note that date format is not used.}

\item{generate_incidence_fun}{a function which takes as arguments age and time and returns a numberic rate of incidence for each age and time included in the simulation.
This function can be defined by user or can be selected from among several default options included in the package.
The user-defined or package default function should be called by name when included as an argument in the generate_incidence_matrix function.}

\item{delta}{the time step between consecurtive birth_dates}
}
\value{
a matrix of column length age_steps and row length birth_dates,
Values stored in the matrix are numeric double, from 0-1, which represent the probability of becoming infected at age and time
}
\description{
a function that returns a matrix of probabilities for each age and time step of the simulation
}
