% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_exclusion_criteria.R
\name{inclusion_exclusion_criteria}
\alias{inclusion_exclusion_criteria}
\title{inclusion_exclusion_criteria}
\usage{
inclusion_exclusion_criteria(required_ages, prevalence, time_step, radii)
}
\arguments{
\item{required_ages}{is a vector of age specific prevalence  at time t1}

\item{prevalence}{is a vector of age specific prevalence  at time t2}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts.}

\item{radii}{inclusion magnitude in years}
}
\value{
a vector or matrice of the specified prevalence or about the age ranges required
}
\description{
a functions that partitions prevalences of given birth cohorts at time 1 and time 2 to m clusters
}
