% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexing_datastructure.R
\name{indexing_datastructure}
\alias{indexing_datastructure}
\title{indexing_datastructure}
\usage{
indexing_datastructure(data_structure, date_of_birth, ages,
  times_since_infection, first_birth_time, last_birth_time, time_step,
  max_age)
}
\arguments{
\item{data_structure}{data sttructure to be indexed}

\item{date_of_birth}{vector of dates of births to be extracted}

\item{ages}{vector of ages to be extracted}

\item{times_since_infection}{vector times since infections to be extracted}

\item{first_birth_time}{date of birth of the first cohort}

\item{last_birth_time}{date of birth of the last cohort}

\item{time_step}{time_step the time step between consecurtive dates or the length of the time between date of births of cohorts}

\item{max_age}{max_age maximum age attained by each birth cohort}
}
\value{
returns the data structure with respect to indexes provided.
}
\description{
a function that takes in a matrix with a pararellogram format and transforms it into a rectangular format and assigns column and row names.
}
