% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulative_probability_surviving_infected.R
\name{infected_cumulative_survival_prob}
\alias{infected_cumulative_survival_prob}
\title{infected_cumulative_survival_prob}
\usage{
infected_cumulative_survival_prob(prob_surviving)
}
\arguments{
\item{prob_surviving}{probability matrix of infected}
}
\value{
returns an array of dimensions time, age and time since infection - tau
}
\description{
a function that returns a matrix of probabilities of mortality for each age and time step of the simulation
}
