% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infected_population.R
\name{infected_population}
\alias{infected_population}
\title{infected_population}
\usage{
infected_population(susceptible, incidence_mat, pmtct_birthcount,
  base_mortality_mat, cumulative_infected_survival, time_step)
}
\arguments{
\item{susceptible}{the susceptible population counts}

\item{incidence_mat}{associated incidence rates}

\item{pmtct_birthcount}{number of newborns infected}

\item{base_mortality_mat}{associated excess mortality survivial rates
the probability of surving infection or death in the susceptible state}

\item{cumulative_infected_survival}{the survival probability of  being aged a at time t having been infected for tau years.}

\item{time_step}{the time or age difference between to consecutive times or ages the probability of surving infection or death in the susceptible statemi.e it is uniform in all values supplied}
}
\value{
returns an array of dimensions time t , age a and time since infection - tau
}
\description{
a function that returns a matrix of probabilities of mortality for each age and time step of the simulation
}
