% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Infected_population_fun.R
\name{infected_population_fun}
\alias{infected_population_fun}
\title{infected_population_fun}
\usage{
infected_population_fun(susceptible_pop_counts, incidence_prob,
  survival_probability)
}
\arguments{
\item{susceptible_pop_counts}{denotes a function that calculates the number of people who where susceptible at a given age and time}

\item{incidence_prob}{denotes a  incidence matrix that yields the probabilities of getting infected at a given age and time.}

\item{survival_probability}{is the survival probability matrix associated with being infected at age a, time t, and for tau years.}
}
\value{
returns an array of dimensions time, age and "time_since_infection - tau"
}
\description{
a function that simulates the infected population from a the susceptibles, incidence and the probalility of surviving to a given age and time having been infected for a time "time_since_infection"
}
\examples{
Infected <- infected_population_fun(susceptible_pop_counts = susceptible_pop_counts,
 incidence_prob = incidence_m,survival_probability = infected_survival_probs)


}
