% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infected_population.R
\name{infected_population_matrix}
\alias{infected_population_matrix}
\title{infected_population_natrix}
\usage{
infected_population_matrix(incidence_mat, base_mortality_mat, time_step,
  susceptible, pmtct_birthcount, cumulative_infected_survival)
}
\arguments{
\item{incidence_mat}{associated incidence rates}

\item{base_mortality_mat}{associated excess mortality survivial rates
the probability of surving infection or death in the susceptible state}

\item{time_step}{the time or age difference between to consecutive times or ages the probability of surving infection or death in the susceptible statemi.e it is uniform in all values supplied}

\item{susceptible}{the susceptible population counts}

\item{pmtct_birthcount}{number of newborns infected}

\item{cumulative_infected_survival}{the survival probability of  being aged a at time t having been infected for tau years.}
}
\value{
returns an matrix of dimensions time t , age a and time since infection - tau
}
\description{
a function that returns a matrix of the infected population for a given birthcohort
}
