% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/look_up_matrice_array.R
\name{look_up}
\alias{look_up}
\title{lookup}
\usage{
look_up(data_structure = matrice, list_of_times = 1985:1990,
  max_age = 3, time_step)
}
\arguments{
\item{data_structure}{the number of steps to age the population}

\item{list_of_times}{a numeric vectors of length min:max; indicates the range of age to be included in simulation. Note that date format is not used.}

\item{max_age}{maximum age each birth cohort is to be aged}

\item{time_step}{the time step between consecurtivelist_of_times}
}
\value{
a data frame with date of birth, age, and current date  if they susceptibles and for infecteds returns a data frame with date of
 birth, age, time since infection, and current date
}
\description{
A wrapper function that returns a a list of the the susceptible and infected population
}
