% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_linear_excess_mortality.R
\name{no_excessmortality}
\alias{no_excessmortality}
\title{no_basemortality}
\usage{
no_excessmortality(ages, times, times_since_i)
}
\arguments{
\item{ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{times}{numeric, indicates time or times at which the incidence rate is desired}

\item{times_since_i}{numeric, indicates age or ages at which the incidence rate is desired}
}
\value{
a numeric vector that represents the mortality rate at t.
}
\description{
a function that takes as arguments age and time and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of mortality at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
