% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PMTCT_birthcounts.R
\name{pmtct_birthcounts}
\alias{pmtct_birthcounts}
\title{PMTCT_birthcounts}
\usage{
pmtct_birthcounts(
  dates_needing_birth_counts,
  pmtct_birth_rate,
  birth_count,
  time_step
)
}
\arguments{
\item{dates_needing_birth_counts}{a vector of births dates to be assigned the birth count}

\item{pmtct_birth_rate}{a function  for the distribution of births in the interval}

\item{birth_count}{total number of births at given date births}

\item{time_step}{the time step between consecurtive birth dates in the vector list_of_birth_times}
}
\value{
Returns a vector of bith counts
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
