% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalences_calculation.R
\name{prevalences_calculation}
\alias{prevalences_calculation}
\title{prevalences_calculation}
\usage{
prevalences_calculation(time_step, recency_type, cohort)
}
\arguments{
\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts.}

\item{recency_type}{type of function to be utilised in the probability of infection function}

\item{cohort}{Status of the popuation in the cohort.}
}
\value{
a list of prevalences i.e. prevalence of HIV status, prevalence of recency,and prevelance of HIV age and tau .
}
\description{
calculates the probability of being recently infected given that you already HIV positive.
}
\details{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
