% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalences_calculation.R
\name{prevalences_calculation_compact}
\alias{prevalences_calculation_compact}
\title{prevalences_calculation_compact}
\usage{
prevalences_calculation_compact(time_step, type, population_at_date,
  probability_of_recent_infection)
}
\arguments{
\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts.}

\item{type}{type of function to be utilised in the probability of infection function}

\item{population_at_date}{Status of the popuation in the population.}

\item{probability_of_recent_infection}{a function that takes in possible  time since infection in years and calculates the probability
of being infected for that amount of years}
}
\value{
a list of prevalences i.e. prevalence of HIV status, prevalence of recency,and prevelance of HIV age and tau .
}
\description{
calculates the probability of being recently infected given that you already HIV positive.
}
\details{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
