% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabiliy_surviving_infected.R
\name{probability_of_surviving_infected}
\alias{probability_of_surviving_infected}
\title{probability_of_surviving_infected}
\usage{
probability_of_surviving_infected(wedge_of_excess_mortality_array,
  matrix_of_base_mortality, time_step)
}
\arguments{
\item{wedge_of_excess_mortality_array}{array containing excess mortality rates for a given time space.}

\item{matrix_of_base_mortality}{matrix containing background rates for a given time space.}

\item{time_step}{the time or age difference between to consecutive times or ages i.e it is uniform in all values supplied}
}
\value{
returns an array of dimensions time, age and time since infection - tau which is the probability of surviving the given time_step given that you are infected.
}
\description{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
