% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabiliy_surviving_infected.R
\name{probability_surviving_infected_matrix}
\alias{probability_surviving_infected_matrix}
\title{probability_of_surviving_infected}
\usage{
probability_surviving_infected_matrix(excess_mortality_matrix, time_step,
  base_mortality_vector)
}
\arguments{
\item{excess_mortality_matrix}{matrix containing excess mortality rates for a given time space.}

\item{time_step}{the time or age difference between to consecutive times or ages i.e it is uniform in all values supplied}

\item{base_mortality_vector}{containing background rates for a given time space.}
}
\value{
returns an matrix of dimensions time, age and time since infection - tau which is the probability of surviving the given time_step given that you are infected.
}
\description{
a function that returns an matrix of probabilities of surviving in the infected state for each age, time and time since infection.
}
