% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_surveydata.R
\name{smoothing_surveydata}
\alias{smoothing_surveydata}
\title{smoothing_surveydata}
\usage{
smoothing_surveydata(age_0, time_0, distance, Data)
}
\arguments{
\item{age_0}{output from the simulation with ages, HIV prevalence, and prevalence of recency.}

\item{time_0}{reporting age bin in a time slice}

\item{distance}{the time difference btween two consecutive times or ages}

\item{Data}{the time difference btween two consecutive times or ages}
}
\value{
individual level data with age, HIV status, and HIV recency status.
}
\description{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
