% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_linear_excess_mortality.R
\name{step_excess_mortality}
\alias{step_excess_mortality}
\title{step_excess_mortality}
\usage{
step_excess_mortality(ages, times, times_since_i)
}
\arguments{
\item{ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{times}{numeric, indicates time or times at which the incidence rate is desired}

\item{times_since_i}{umeric, indicates time since infection among the infected poplation.}
}
\description{
a function that takes as arguments age,time and tau - which indicates the average time since infection among the infected population - and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of excess mortality -i.e. among infected population relative to non-infected population - at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
