% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_indep_age_linear_base_mortality.R
\name{step_mortality}
\alias{step_mortality}
\title{step_mortality}
\usage{
step_mortality(ages, times)
}
\arguments{
\item{ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{times}{numeric, indicates time or times at which the incidence rate is desired}
}
\value{
a numeric vector that represents the mortality rate at t.
}
\description{
a function that takes as arguments age and time and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of mortality at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
