% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pps_survey_selfweighting_structured.R
\name{Survey_pps}
\alias{Survey_pps}
\title{Survey_pps}
\usage{
Survey_pps(cluster_number, cluster_size, num_cluster_sample,
  ind_per_cluster, overall_prevalence1, overall_prevalence2,
  sigma_prevalence)
}
\arguments{
\item{cluster_number}{unique id number of the cluster}

\item{cluster_size}{sizes of the clusters}

\item{num_cluster_sample}{the number of people in the cluster}

\item{ind_per_cluster}{individuals to sample per cluster}

\item{overall_prevalence1}{is a vector of age specific prevalence  at time t1}

\item{overall_prevalence2}{is a vector of age specific prevalence  at time t2}

\item{sigma_prevalence}{is a vector with the  standard errors associated with age specific prevalence  at time t1}
}
\value{
returns an array of dimensions time t , age a and time since infection - tau
}
\description{
conducts surveys using Probability propotional to size (pps) in given regions.
}
