% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptible_cumulative_survival_fun.R
\name{susceptible_cumulative_survival_fun}
\alias{susceptible_cumulative_survival_fun}
\title{generate_susceptible_survival}
\usage{
susceptible_cumulative_survival_fun(incidence_matrix,
  base_mortality_matrix, delta)
}
\arguments{
\item{incidence_matrix}{probability matrix of getting infected in the susceptible}

\item{base_mortality_matrix}{probability matrix of dying among the susceptible}

\item{delta}{the time or age difference between consecutive ages or time and it is uniform}
}
\value{
a matrix calculated from the susceptible_survival_rate_matrix
Values stored in the matrix are numeric-double, from 0-1, which represent the  probability of surviving a specified age from birth in the susceptible population.
}
\description{
a function that returns a matrix of cumulative probabilities of survival -i.e. not dying and aquiring infection - for each age and time step of the simulation
}
