% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptible_cumulative_survival.R
\name{susceptible_cumulative_survival_vector}
\alias{susceptible_cumulative_survival_vector}
\title{susceptible_cumulative_survival_vector}
\usage{
susceptible_cumulative_survival_vector(incidence_vector,
  base_mortality_vector, time_step)
}
\arguments{
\item{incidence_vector}{vector of  incidence rates for specified ages and times}

\item{base_mortality_vector}{vector of  natural mortality rates for specified ages and times}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts}
}
\value{
a vector of cumulative probabilities (from birth till age a at time t) of surviving in the susceptible population without being infected or dying from natural death  .
Values stored in the matrix are numeric-double, from 0-1, which represent
the  probability of surviving a specified age from birth in the susceptible population.
}
\description{
a function that returns a matrix of cumulative probabilities of survival -i.e. not dying and aquiring infection - for each age and time step of the simulation
}
