% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptibles_population.R
\name{susceptible_population}
\alias{susceptible_population}
\title{susceptible_population}
\usage{
susceptible_population(cumulative_survival_matrix, birth_counts)
}
\arguments{
\item{cumulative_survival_matrix}{matrix of cumulative survival probabilities in the susceptible from birth till the specified age and time.}

\item{birth_counts}{vector of numberic values representing the total birth counts for each birth count..}
}
\value{
a matrix of the susceptible population at a specified age and time.
number in each cell of the matrix represent the number of individuals in the population who are alive and un-infected
}
\description{
a function that returns a matrix of numeric values for age and time, representing the number of individuals within the population who are alive and not infected
}
