% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_linear_excess_mortality.R
\name{tau_linear_excess_mortality}
\alias{tau_linear_excess_mortality}
\title{tau_linear_excess_mortality}
\usage{
tau_linear_excess_mortality(ages, times, times_since_i, intercept = 0.05,
  slope = 0)
}
\arguments{
\item{ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{times}{numeric, indicates time or times at which the incidence rate is desired}

\item{times_since_i}{numeric, indicates a constant rate of mortality when}

\item{intercept}{numeric, indicates minimum age to be included in the simulation}

\item{slope}{numeric, indicates maximum age to be included in the simulation}
}
\value{
a numeric vector that represents the mortality rate at t.
}
\description{
a function that takes as arguments age,time and tau (which indicates the average time since infection among the infected population)
and returns a numeric vector with length equivelent to the number of times indicated by the simulation
representing the excess mortality rate  induced by being infected
}
\details{
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
