% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_indep_age_linear_base_mortality.R
\name{time_indep_age_linear_base_mortality}
\alias{time_indep_age_linear_base_mortality}
\title{time_indep_age_linear_base_mortality}
\usage{
time_indep_age_linear_base_mortality(ages, times, constant = 0,
  age_min = 0, age_max = 50, mort_min = 0.01, mort_max = 0.03)
}
\arguments{
\item{ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{times}{numeric, indicates time or times at which the incidence rate is desired}

\item{constant}{numeric, indicates a constant rate of mortality when}

\item{age_min}{numeric, indicates minimum age to be included in the simulation}

\item{age_max}{numeric, indicates maximum age to be included in the simulation}

\item{mort_min}{numeric, indicates minimum mortality, which is at age_min}

\item{mort_max}{numeric, indicates maximum/final mortality at age_max, unless otherwise specified by user defined function}
}
\value{
a numeric vector that represents the mortality rate at t.
}
\description{
a function that takes as arguments age and time and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of mortality at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
