% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_indep_age_linear_excess_mortality.R
\name{time_indep_age_linear_excess_mortality}
\alias{time_indep_age_linear_excess_mortality}
\title{time_indep_age_linear_excess_mortality}
\usage{
time_indep_age_linear_excess_mortality(matrix_of_ages, matrix_of_times,
  constant = 0, age_min = 0, times_since_i, age_max = 50,
  mort_min = 0.01, mort_max = 0.05)
}
\arguments{
\item{matrix_of_ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{matrix_of_times}{numeric, indicates time or times at which the incidence rate is desired}

\item{constant}{numeric, indicates a constant rate of mortality when}

\item{age_min}{numeric, indicates minimum age to be included in the simulation}

\item{times_since_i}{umeric, indicates time since infection among the infected poplation}

\item{age_max}{numeric, indicates maximum age to be included in the simulation}

\item{mort_min}{numeric, indicates minimum mortality, which is at age_min}

\item{mort_max}{numeric, indicates maximum/final mortality at age_max, unless otherwise specified by user defined function}
}
\value{
a numeric vector that represents the mortality rate at t.
}
\description{
a function that takes as arguments age,time and tau - which indicates the average time since infection among the infected population - and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of excess mortality -i.e. among infected population relative to non-infected population - at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
