% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data_fun.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(data_to_transform)
}
\arguments{
\item{data_to_transform}{the matrix to be transformed
The user-defined or package default function should be called by name when included as an argument in the generate_base_mortality_matrix function.}
}
\value{
returns a matrix with a rectangular format, labelled rows and columns
}
\description{
a function that takes in a matrix with a pararellogram format and transforms it into a rectangular format and assigns column and row names.
}
