% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_linear_excess_mortality.R
\name{weibull_hazard}
\alias{weibull_hazard}
\title{weibull_hazard}
\usage{
weibull_hazard(matrix_of_ages, matrix_of_times, times_since_i, shape = 2,
  scale = 5, base = T, ...)
}
\arguments{
\item{matrix_of_ages}{numeric, indicates age or ages at which the incidence rate is desired}

\item{matrix_of_times}{numeric, indicates time or times at which the incidence rate is desired}

\item{times_since_i}{umeric, indicates time since infection among the infected poplation.}

\item{shape}{numeric, indicates minimum age to be included in the simulation}

\item{scale}{numeric, indicates minimum mortality, which is at age_min}

\item{base}{allows generic use as a hazard function excess mort or base mortality}

\item{...}{extra parameters}
}
\value{
a numeric vector that represents the mortality rate at
}
\description{
a function that takes as arguments age,time and tau - which indicates the average time since infection among the infected population - and returns a numeric vector of length equivelent to the number of times indicated by the simulation
representing a rate of excess mortality -i.e. among infected population relative to non-infected population - at the indicated age and time
The generate_mortality function is required as an argument for the package's do_simulation function
The function may be user defined and stored as an R object. Otherwise a default value - entered as "default" - is provided by the package
}
