% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_mortality_matrix.R
\name{base_mortality_matrix}
\alias{base_mortality_matrix}
\title{base_mortality_matrix_fun}
\usage{
base_mortality_matrix(max_age, list_of_times, base_mortality, time_step)
}
\arguments{
\item{max_age}{a number. Indicates the number of steps forward each age group will be aged in the simulation by the do_sim function}

\item{list_of_times}{a numeric vector of length min:max; indicates the range ofage to be included in simulation. Note that date format is not used.}

\item{base_mortality}{a function which takes as arguments age and time and returns a numberic rate of mortality for each age and time included in the simulation.
This function can be defined by user or can be selected from among several default options included in the package.
The user-defined or package default function should be called by name when included as an argument in the generate_base_mortality_matrix function}

\item{time_step}{the time step between consecurtive list_of_times}
}
\value{
a matrix of column length max_age and row length list_of_times.
Values stored in the matrix are numeric-double, from 0-1, which represent the probability of dying at each age and time
}
\description{
a function that returns a matrix of probabilities of mortality for each age and time step of the simulation
}
