% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_counts_fun.R
\name{birth_counts_fun}
\alias{birth_counts_fun}
\title{birth_counts_fun}
\usage{
birth_counts_fun(total_births, birth_dates, delta, birthrate)
}
\arguments{
\item{total_births}{the total number of births between maximum and minimun birth_dates given.}

\item{birth_dates}{the calender dates of births.}

\item{delta}{the time step between consecurtive birth_dates}

\item{birthrate}{a function  for the distribution of births in interval}
}
\value{
Returns a vector of bith counts  from the inputs total_births, delta and birth_dates of length min-birth_dates:max-birth_dates
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
