% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birth_rate_fun.R
\name{birth_rate_fun}
\alias{birth_rate_fun}
\title{birth_rate_fun}
\usage{
birth_rate_fun(tiam)
}
\arguments{
\item{tiam}{is a double/vector of times the birth rate is to be evaluated at.}
}
\value{
Returns a vector of bith counts  from the inputs total_births, delta and birth_dates of length min-birth_dates:max-birth_dates
}
\description{
a function that returns a numeric vector of birth counts from the arguments of total births, a range of birth dates and an indicted birth rate
}
