% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extactcohorttau_distribution.R
\name{estimate_excessmortality}
\alias{estimate_excessmortality}
\title{calculate_excessmortality}
\usage{
estimate_excessmortality(populationdistribution, excess_mortality,
  base_mortality, reporting_bin)
}
\arguments{
\item{populationdistribution}{the cohort distribution of taus since infection}

\item{excess_mortality}{excess mortality resulting from the infection}

\item{base_mortality}{natural causes of death}

\item{reporting_bin}{the size of the reporting bin default is 1}
}
\value{
a data frame of ages, times and the corresponding excess mortality rate
}
\description{
A wrapper function that returns a list of the the susceptible and infected population
}
