% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cohort_status_surveydate.R
\name{extract_cohort_status_surveydate}
\alias{extract_cohort_status_surveydate}
\title{extract_cohort_status_surveydate}
\usage{
extract_cohort_status_surveydate(date_of_birth, max_age, time_slice,
  time_step, population)
}
\arguments{
\item{date_of_birth}{the date of birth of the cohort.}

\item{max_age}{maximum age of the birth cohort.}

\item{time_slice}{dates with which the survey are to be conducted}

\item{time_step}{the time or age difference between to consecutive times or ages the probability of surving infection or
death in the susceptible statemi.e it is uniform in all values supplied}

\item{population}{dates with which the survey are to be conducted}
}
\value{
a vector which denotes being the probability of testing recently infected  being recently infected fo
}
\description{
a function that calculates the probability of testing recently infected at a time since infection tau.
}
