% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infected_population.R
\name{infected_population}
\alias{infected_population}
\title{infected_population}
\usage{
infected_population(susceptible = susceptible_pop_counts[,
  -ncol(susceptible_pop_counts)], incidence_matrix_mod = mod_incidence,
  cumulative_infected_survival = Cum_prob_survival_i)
}
\arguments{
\item{susceptible}{the susceptible population counts}

\item{incidence_matrix_mod}{the normalised incidence with respect to the exponetial used in approximating the probability of surving infection or death in the susceptible state}

\item{cumulative_infected_survival}{the survival probability of  being aged a at time t having been infected for tau years.}
}
\value{
returns an array of dimensions time t , age a and time since infection - tau
}
\description{
a function that returns a matrix of probabilities of mortality for each age and time step of the simulation
}
