% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_recent_among_positive.R
\name{probability_recent_among_positive}
\alias{probability_recent_among_positive}
\title{probability_recent_among_positive}
\usage{
probability_recent_among_positive(time_step, population_at_date)
}
\arguments{
\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts.}

\item{population_at_date}{Status of the popuation in the population.}
}
\value{
bl bl bl
}
\description{
Takes in any numeric value and squares it
}
\details{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
