% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susceptibles_population.R
\name{susceptible_population}
\alias{susceptible_population}
\title{susceptible_fun}
\usage{
susceptible_population(cumulative_survival_matrix, birth_counts)
}
\arguments{
\item{cumulative_survival_matrix}{matrix of survival probabilities in the susceptible for each age and time.}

\item{birth_counts}{vector of numberic values representing the number of  births at each time between time_1 and time_2.}
}
\value{
a matrix of the susceptible at specified ages and calculations. Numeric values in each cell of the matrix represent the number of individuals in the population who are alive and not infected
}
\description{
a function that returns a matrix of numeric values for age and time, representing the number of individuals within the population who are alive and not infected
}
