% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suveryimplemetation.R
\name{suveryimplemetation}
\alias{suveryimplemetation}
\title{suveryimplemetation}
\usage{
suveryimplemetation(survey_status, prevalencesdata, time_step)
}
\arguments{
\item{survey_status}{demographic specifications of the survey.}

\item{prevalencesdata}{output from the simulation with ages, HIV prevalence, and prevalence of recency.}

\item{time_step}{the time step between consecurtive dates or the length of the time between date of births of cohorts}
}
\value{
individual level data with age, HIV status, and HIV recency status.
}
\description{
a function that returns an array of probabilities of surviving in the infected state for each age, time and time since infection.
}
