% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data_fun.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(age_max, data_to_transform, date_of_birth, delta = 1)
}
\arguments{
\item{age_max}{denotes the maximum age reached by each cohort.}

\item{data_to_transform}{the matrix to be transformed}

\item{date_of_birth}{a vector with two numbers giving the minimum and maximum calender dates.}

\item{delta}{the time difference between consecutive dates of birth (if not specified defaults to 1).
The user-defined or package default function should be called by name when included as an argument in the generate_base_mortality_matrix function.}
}
\value{
returns a matrix with a rectangular format, labelled rows and columns
}
\description{
a function that takes in a matrix with a pararellogram format and transforms it into a rectangular format and assigns column and row names.
}
