% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wedge_excess_mortality_array.R
\name{wedge_excess_mortality_array}
\alias{wedge_excess_mortality_array}
\title{wedge_excess_mortality_array}
\usage{
wedge_excess_mortality_array(max_age, list_of_birth_times,
  excess_mortality, time_step)
}
\arguments{
\item{max_age}{the optomal age each birth cohort is aged to}

\item{list_of_birth_times}{the birth dates of each of the birth cohort}

\item{excess_mortality}{the excess mortality function}

\item{time_step}{the time difference between two consective times}
}
\value{
returns an array of dimensions time, age and time since infection - tau, which is essentially the probability of surviving in the given time space.
}
\description{
a function that returns a matrix of probabilities of mortality for each age and time step of the simulation
}
